Fail the stable release job on curl failures
curl's default behavior is to succeed even when the request have
failed. This means that if any error happens whil triggering
the charts.gitlab.io release pipeline, the job in GitLab Runner Helm
Chart project still succeeds. That makes us unaware of the release
not being finished.

There is a `--fail` option that changes that, but it also prevents
printing to the output. So with it we could know that some problem
have happened, but we would not know what the problem was.

In version 7.76, at February 2021, a new option `--fail-with-body` was
added. And it seems that it will resolve all of our problems. For that
we would however need a newer environment than the `ruby:2.6.5-alpine`
image (released more than 2 years ago) used as base of
`registry.gitlab.com/gitlab-org/gitlab-build-images:gitlab-charts-build-base`.

But as the only command that we execute for the stable release job is
the curl command, there is no need to use the charts specific image.
We can defer to base `alpine` in the newest version and the newest
curl version available there.

Which is what this commit does.